WIP: Refactoring single and multi-select into 2 components.
