+++ e2e/mocks/profile-api/demo.json
@@ -70,10 +70,7 @@
       "key": "DataApiOptionMultiSelect",
       "label": "DataApiOption with filter multi select",
       "required": false,
-      "displayType": "filter-select",
-      "filterSelectConfig": {
-        "enableMultipleSelection": true
-      },
+      "displayType": "multi-select",
       "dynamicSelectConfig": {
         "type": "test-type",
         "labelPath": "label"
