+++ src/app/content/content-metadata/content-metadata.component.html
@@ -4,7 +4,6 @@
       <li *ngFor="let fieldToDisplay of pageConfig.fieldsToDisplay; index as i">
         <div [ngSwitch]="fieldToDisplay.displayType">
           <app-options-autocomplete *ngSwitchCase="'filter-select'"
-            [multiSelect]="fieldToDisplay.filterSelectConfig && fieldToDisplay.filterSelectConfig.enableMultipleSelection"
             [formControlName]="fieldToDisplay.key"
             [placeholder]="fieldToDisplay.label"
             [fieldConfig]="fieldToDisplay"
@@ -14,6 +13,16 @@
             [id]="'custom-input-'+ i"
             [attr.id]="'custom-input-'+ i">
           </app-options-autocomplete>
+          <app-options-multiselect *ngSwitchCase="'multi-select'"
+            [formControlName]="fieldToDisplay.key"
+            [placeholder]="fieldToDisplay.label"
+            [fieldConfig]="fieldToDisplay"
+            [parentControl]="formGroup.get('metadata').get(fieldToDisplay.dynamicSelectConfig?.parentFieldConfig?.key || '')"
+            [errorMessage]="getErrorMessage(fieldToDisplay.key)"
+            [required]="fieldHasRequiredError(fieldToDisplay)"
+            [id]="'custom-input-'+ i"
+            [attr.id]="'custom-input-'+ i">
+          </app-options-multiselect>
           <mat-form-field *ngSwitchCase="'select'">
             <app-options-input
               [formControlName]="fieldToDisplay.key"
