--- src/app/shared/widgets/options-autocomplete/options-autocomplete.component.css
@@ -1,3 +1,0 @@
-mat-chip mat-icon.remove-option {
-  opacity: 0.7;
-}
