+++ src/app/shared/widgets/options-autocomplete/options-autocomplete.component.html
@@ -1,5 +1,5 @@
 <mat-form-field [formGroup]="formGroup">
-  <input type="text" *ngIf="!multiSelect"
+  <input type="text"
       matInput
       formControlName="internalFilterControl"
       [matAutocomplete]="auto"
@@ -9,32 +9,11 @@
       [attr.aria-disabled]="disabled"
       [attr.aria-required]="required" />
 
-  <mat-chip-list #chipList *ngIf="multiSelect"
-      aria-label="Options selection"
-      [required]="required"
-      [disabled]="disabled"
-      [attr.aria-disabled]="disabled"
-      [attr.aria-required]="required">
-    <mat-chip *ngFor="let option of selectedOptions"
-              (removed)="removeOptionFromMultiSelect(option)">
-      {{option.displayValue}}
-      <mat-icon matChipRemove class="remove-option">cancel</mat-icon>
-    </mat-chip>
-    <input
-      #filterInputElement
-      formControlName="internalFilterControl"
-      [placeholder]="multiSelectLabelText"
-      [attr.aria-label]="placeholder"
-      [matAutocomplete]="auto"
-      [matChipInputFor]="chipList" />
-  </mat-chip-list>
-
   <mat-autocomplete #auto="matAutocomplete"
                     [displayWith]="getOptionDisplayValue"
                     (closed)="optionListClosed()"
-                    (opened)="announceWithSelectionCount()"
                     (optionSelected)="optionSelected($event.option.value)">
-    <mat-option *ngFor="let option of filteredOptions$ | async" [value]="option" [disabled]="!areMoreSelectionsAllowed">
+    <mat-option *ngFor="let option of filteredOptions$ | async" [value]="option">
       {{ option.displayValue }}
     </mat-option>
   </mat-autocomplete>
