+++ src/app/shared/widgets/options-multiselect/options-multiselect.component.css
@@ -1,0 +1,3 @@
+mat-chip mat-icon.remove-option {
+  opacity: 0.7;
+}
