+++ src/app/shared/widgets/options-multiselect/options-multiselect.component.html
@@ -1,0 +1,34 @@
+<mat-form-field [formGroup]="formGroup">
+  <mat-chip-list #chipList
+      aria-label="Options selection"
+      [required]="required"
+      [disabled]="disabled"
+      [attr.aria-disabled]="disabled"
+      [attr.aria-required]="required">
+    <mat-chip *ngFor="let option of selectedOptions"
+              (removed)="removeOptionFromMultiSelect(option)">
+      {{option.displayValue}}
+      <mat-icon matChipRemove class="remove-option">cancel</mat-icon>
+    </mat-chip>
+    <input
+      #filterInputElement
+      formControlName="internalFilterControl"
+      [placeholder]="multiSelectLabelText"
+      [attr.aria-label]="placeholder"
+      [matAutocomplete]="auto"
+      [matChipInputFor]="chipList" />
+  </mat-chip-list>
+
+  <mat-autocomplete #auto="matAutocomplete"
+                    [displayWith]="getOptionDisplayValue"
+                    (opened)="announceWithSelectionCount()"
+                    (optionSelected)="addOptionToMultiSelect($event.option.value)">
+    <mat-option *ngFor="let option of filteredOptions$ | async" [value]="option" [disabled]="!areMoreSelectionsAllowed">
+      {{ option.displayValue }}
+    </mat-option>
+  </mat-autocomplete>
+  <mat-icon *ngIf="disabled" matSuffix class="disabled-icon" matTooltip="Read-only">lock</mat-icon>
+  <mat-error>
+    {{errorMessage}}
+  </mat-error>
+</mat-form-field>
